requirements.txt
Django==5.1.2
celery==5.4.0
sentry-sdk==2.1.1
structlog==24.4.0
django-environ==0.11.2
pydantic==2.9.2
pytest==8.3.3
pytest-django==4.9.0
psycopg2==2.9.10
ruff==0.7.1
clickhouse-connect==0.8.5
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-instrumentation-django==0.49b1
opentelemetry-instrumentation-celery==0.49b1
opentelemetry-exporter-otlp==1.28.1
redis
